+++ iterator/ts/aggregator.ts
@@ -1,0 +1,6 @@
+import {Iterator} from "./iterator";
+
+export interface Aggregator {
+    // Get an external iterator.
+    getIterator(): Iterator<any>;
+}
